app/filesystem_store.go
package main

import (
	"encoding/json"
	"io"
)

type FileSystemPlayerStore struct {
	DB     io.ReadWriteSeeker
	league League
}

func NewFileSystemPlayerStore(db io.ReadWriteSeeker) *FileSystemPlayerStore {
	db.Seek(0, 0)
	league, _ := NewLeague(db)
	return &FileSystemPlayerStore{
		db, league,
	}
}

func (f *FileSystemPlayerStore) GetLeague() League {
	return f.league
}

func (f *FileSystemPlayerStore) GetPlayerScore(name string) int {
	player := f.league.Find(name)

	if player != nil {
		return player.Score
	}

	return 0
}

func (f *FileSystemPlayerStore) PostPlayerScore(name string) {

	player := f.league.Find(name)

	if player != nil {
		player.Score++
	} else {
		f.league = append(f.league, Player{name, 1})
	}

	f.DB.Seek(0, 0)
	json.NewEncoder(f.DB).Encode(f.league)
}
